refs CRASM-3382: Move hook tests and complete Jest -> Vitest migration (Step 3/8)
- Move 3 hook test files to tests/hooks/ with .test extension
- Convert all Jest APIs to Vitest equivalents (spyOn, mocked, timers, etc.)
- Replace jest-date-mock with native vi.setSystemTime()
- Update import paths and remove empty directories
- Verified: functional tests pass after migration
- TODO for pre-existing useVulnScanData mocking issues
