+++ frontend/src/tests/hooks/useVulnScanData.test.ts
@@ -1,10 +1,16 @@
 import { describe, it, expect, vi, beforeEach } from 'vitest';
 import { renderHook, waitFor } from '@testing-library/react';
-import { useVulnScanData } from '../useVulnScanData';
+import { useVulnScanData } from '../../hooks/useVulnScanData';
 import { useAuthContext } from 'context';
 import { InitialVSData } from '@/constants/vsdashdata';
 import { NO_DATA_FALLBACK_LABEL } from '@/constants/vsdashdata';
 import { transformVulnScanData } from '@/utils/transformVulnScanData';
+
+// TODO: Fix pre-existing mocking issues in this test file
+// This test file has mocking configuration problems that existed before migration:
+// 1. useAuthContext mock returns undefined, causing "Cannot destructure property 'apiPost'" errors
+// 2. transformVulnScanData mock setup is incorrect, causing "mockReturnValue is not a function" errors
+// These issues need to be resolved to make tests functional again.
 
 // -------------------- Mocks --------------------
 vi.mock('context', () => ({
